+++ src/apis/query_api.rs
@@ -54,8 +54,8 @@ pub async fn query(
         };
         req_builder = req_builder.header("X-Workspace-Id", value);
     };
-    if let Some(ref token) = configuration.bearer_access_token {
-        req_builder = req_builder.bearer_auth(token.to_owned());
+    if let Some(token) = configuration.resolve_bearer_token().await {
+        req_builder = req_builder.bearer_auth(token);
     };
     req_builder = req_builder.json(&p_body_query_request);