+++ src/apis/refresh_api.rs
@@ -47,8 +47,8 @@ pub async fn refresh(
         };
         req_builder = req_builder.header("X-Workspace-Id", value);
     };
-    if let Some(ref token) = configuration.bearer_access_token {
-        req_builder = req_builder.bearer_auth(token.to_owned());
+    if let Some(token) = configuration.resolve_bearer_token().await {
+        req_builder = req_builder.bearer_auth(token);
     };
     req_builder = req_builder.json(&p_body_refresh_request);